Move Set std to SetInstances
